baikal: Add 8GB memory layout support
Baikal-T1 SoC being 32-bit P5600 MIPS Warrior based chip supports
XPA - eXtended Physical Addressing, which provides a way to
address more than 4Gb of memory to the user-space. The corresponding
functionality is already developed in kernel and memory SPD blob
parser, but several bugs prevent system to support more than 4GB
of memory.

First of all even if DDR is initialized with proper SDRAM layout a
value passed to kernel can't exceed 0xDFFFFFFF0 due to the constant
MAX_HIGHMEM_SIZE. We need to increase it value to 0x1DFFFFFF0 so
to comply with the 8GB limitation.

Secondly the kernel dts scripts limit the CPU memory address and
size cells to being just 1 each, which means both address and
size can be 32-bit wide, which isn't enough to pass 8GB memory
ranges. So we need to alter the code responsible for setting up
the memory-related data into the kernel fdt taking into account
the memory node addr and size cells. The rest should be done in
the kernel dts files.

+++ include/board.h
@@ -8,6 +8,6 @@
 #ifndef _BOARD_H_
 #define _BOARD_H_
 
-uint32_t get_ddr_highmem_size(void);
+uint64_t get_ddr_highmem_size(void);
 
 #endif /* _BOARD_H_ */
